requirements.txt
pandas==2.1.0
pymongo==4.5.0
boto3==1.28.50
boto3_type_annotations==0.3.1
overrides==7.4.0
pytest==7.4.2
python-dotenv==1.0.0
marshmallow==3.20.1
pre-commit==3.4.0
dependency-injector==4.41.0
